modularizes styles and classnames
