+++ src/scripts/modules/components/react/components/NewComponentSelection.less
@@ -2,15 +2,6 @@
   float: none;
 }
 
-.kbc-components-list .kbc-header {
-  position: relative;
-
-  .badge-component-container-plain {
-    top: 24px;
-    right: 24px;
-  }
-}
-
 .kbc-components-overview .td {
   position: relative;
   padding: 24px;
