+++ src/scripts/react/common/ComponentBadgeCell.jsx
@@ -1,7 +1,5 @@
 import React, {PropTypes} from 'react';
 import ComponentType from './ComponentType';
-import { Icon } from '@keboola/indigo-ui';
-import { ExternalLink } from '@keboola/indigo-ui';
 
 require('./Badges.less');
 
@@ -14,7 +12,7 @@ export default React.createClass({
     const badges = this.getBadges();
 
     return (
-      <div className="badge-component-container-plain">
+      <div className="badge-component-container badge-component-container-selection">
         {badges.map((badge, idx) =>
           <div className={'badge badge-component-item badge-component-item-' + badge.key}
             title={badge.description}
@@ -26,45 +24,15 @@ export default React.createClass({
       </div>
     );
   },
-
   getBadges() {
-    const flags = this.resolveFlags();
+    const flags = this.props.component.get('flags');
     const componentType = ComponentType.getComponentType(this.props.component.get('type'));
     let badges = [];
-
-    if (!flags.contains('3rdParty')) {
-      badges.push({
-        title: <span><Icon.Keboola className="badge-component-item-responsibility-icon" /> Keboola</span>,
-        description: `Support for this ${componentType} is provided by Keboola.`,
-        key: 'responsibility'
-      });
-    }
     if (flags.contains('3rdParty')) {
       badges.push({
         title: <span>3<sup>rd</sup> party</span>,
         description: `This is a third-party ${componentType} supported by its vendor.`,
         key: '3rdParty'
-      });
-    }
-    if (flags.contains('excludeFromNewList')) {
-      badges.push({
-        title: 'Alpha',
-        description: `This ${componentType} is private.`,
-        key: 'excludeFromNewList'
-      });
-    }
-    if (flags.contains('appInfo.dataIn')) {
-      badges.push({
-        title: <span><i className="fa fa-cloud-download fa-fw"/> IN</span>,
-        description: `This ${componentType} retrieves data from outside sources.`,
-        key: 'dataIn'
-      });
-    }
-    if (flags.contains('appInfo.dataOut')) {
-      badges.push({
-        title: <span><i className="fa fa-cloud-upload fa-fw"/> OUT</span>,
-        description: `This ${componentType} sends data outside of Keboola Connection.`,
-        key: 'dataOut'
       });
     }
     if (flags.contains('appInfo.beta')) {
@@ -74,49 +42,6 @@ export default React.createClass({
         key: 'appInfo.beta'
       });
     }
-    if (flags.contains('appInfo.fee')) {
-      badges.push({
-        title: <span><i className="fa fa-dollar fa-fw"/></span>,
-        description: `There is an extra charge to use this ${componentType}.`,
-        key: 'fee'
-      });
-    }
-    if (flags.contains('appInfo.redshiftOnly')) {
-      badges.push({
-        title: <span><i className="fa fa-database fa-fw"/></span>,
-        description: `A Redshift backend is required to use this ${componentType}.`,
-        key: 'redshift'
-      });
-    }
-    if (flags.contains('appInfo.fullAccess')) {
-      badges.push({
-        title: <span><i className="fa fa-key fa-fw"/></span>,
-        description: `This ${componentType} will have full access to the project including all its data.`,
-        key: 'fullAccess'
-      });
-    }
-    if (flags.contains('deprecated')) {
-      badges.push({
-        title: <span><i className="fa fa-exclamation-triangle fa-fw"/><i className="fa fa-clock-o fa-fw"/></span>,
-        description: `This ${componentType} is deprecated.`,
-        key: 'deprecated'
-      });
-    }
-    if (flags.contains('hasLicence')) {
-      badges.push({
-        title: <span><i className="fa fa-file-text-o fa-fw"/></span>,
-        description: <span>You agree to the <ExternalLink href={this.props.component.getIn(['data', 'vendor', 'licenseUrl'])}>vendor's license agreement</ExternalLink>.</span>,
-        key: 'license'
-      });
-    }
     return badges;
-  },
-
-  resolveFlags() {
-    if (this.props.component.getIn(['data', 'vendor', 'licenseUrl'])) {
-      return this.props.component.get('flags').push('hasLicence');
-    } else {
-      return this.props.component.get('flags');
-    }
   }
 });
